refactor user types
